pmic_lld: IRQ Module misrac fixes rule-10.3
1.IRQ Module misrac fixes rule-10.3 - MISRA.ETYPE.CATEGORY.DIFFERENT.2012
  The operands of the operator == do not have the same essential type category:
  one is unsigned and the other is Boolean.

2.IRQ Module misrac fixes rule-10.3 - MISRA.ETYPE.ASSIGN.2012
  An expression value of essential type unsigned int is assigned to an
  object of essential type unsigned short int

+++ src/cfg/lp8764x/pmic_irq_lp8764x.c
@@ -1039,7 +1039,7 @@ static int32_t Pmic_lp8764x_getFSMErr(Pmic_CoreHandle_t *pPmicCoreHandle,
  * \brief  Function to decipher the L2 Error for LP8764x Hera PMIC
  */
 int32_t Pmic_lp8764x_irqGetL2Error(Pmic_CoreHandle_t *pPmicCoreHandle,
-                                   uint32_t           l1RegAddr,
+                                   uint16_t           l1RegAddr,
                                    Pmic_IrqStatus_t  *pErrStat)
 {
     int32_t pmicStatus  = PMIC_ST_SUCCESS;
