+++ src/cfg/tps6594x/pmic_irq_tps6594x_priv.h
@@ -254,7 +254,7 @@ void pmic_get_tps6594x_intrGpioCfg(Pmic_GpioIntrTypeCfg_t **pIntGpioCfg);
  * \brief  Function to decipher the L2 Error for TPS6594x Leo PMIC.
  */
 int32_t Pmic_tps6594x_irqGetL2Error(Pmic_CoreHandle_t *pPmicCoreHandle,
-                                    uint32_t           l1RegAddr,
+                                    uint16_t           l1RegAddr,
                                     Pmic_IrqStatus_t  *pErrStat);
 
 #ifdef __cplusplus
